Remove unnecessary reconfiguration before daemonizing
During startup, the configuration for log.screen,
engine.autoreload.on, and request.show_tracebacks was overridden to
ensure sane values. But this makes the corresponding entries in the
conf file useless.

Instead, give the conf file authority for these values. But do force
the value of tools.encode.on, which doesn't otherwise make sense as an
editable value.