Use absolute paths in default.conf for database
For unknown reason, the init script does not chdir to the application
directory as expected, creating problems with endpoints that assume
the cwd is the app directory. Avoid this assumption by using an
absolute path, and make the db a sibiling of the app directory rather
than a child.

Do the same for the access and error logs.