Populate page title from later bookmarklet and simplify bookmarking
The bookmarklet now passes in document.title so that medley doesn't
have to do a url fetch until the bookmarklet is saved. The bookmarklet
will also populate the comment field with the value of the meta
descriptiont tag if available and if there is no selected text.

On the server, skip the cache lookup. It made sense when there was a
preliminary fetch to extract the page title so that the subsequent
save would not refetch. Now there is only one fetch.