Migrate headers to an app
Rather than being defined within the medley server, headers is now an
app that gets mounted by the server. The eventual goal is to treat all
exsting endpoints like this, in order to decrease the size of the main
server file and to leverage restful route dispatch when it would be
useful.