[bounce] Associate URLs with names, and finish intend functionality
Name and url get stored in the registry as newline-separated values.

The add form isn't shown when the inbound URL belongs to a known site,
and that site is also excluded from the list of available destinations.

The page will try to guess the site name and group based on some basic
heuristics.