[geodb] Use GET to trigger database download
POST was previously used with an action parameter, but this parameter
was then removed since it only ever has one value. This resulted in a
body-less POST, which is weird.

Shift the download operation back to GET and trigger from a
subpath. This preserves the existing check that prevents unnecessary
redownloads.