Restructure testing setup
Add a test target to the Makefile which will run the tests for each app.

To help pytest identify test files, the test files are renamed. Pytest
is invoked through python so that PYTHONPATH does not need to be
accounted for.

Many tests have been marked as skipped pending further refactoring.