[headers] Introduce a new approach to content negotiation and templating
In an effort to reduce repetition across controller classes, combine
the negotiable and template tools.

The controller no longer needs to consider the format of the
response. It leaves that to the negotiable tool, which decides on the
most acceptable response type as before but also applies
templates. This makes less decorator stacking in the controller.

Within the test suite, switch to a mixin approach for custom
assertions to accomplish similar goals of decreasing redundancy.

Existing apps will be upgraded to this new setup individually.