Switch virtualenv creation to use built-in module
Tested against Python 3.5. A custom prompt is no longer used.

Added a reminder to source the activate script. This has to happen
separately so that it affects the login shell.

Added --system-site-packages in spite of previous disallow it. There
may have been historical reasons for doing it that way, but now it
only makes things take longer and has no strong benefit.