Add multi method, don't return id from insert
The insert method cannot return the lastrowid because that value is
not populated due to the use of executemany(). It also would not be
populated if the query is anything other than a standard insert.

The multi method provides an alternative for being less dependent on
lastrowid. Taking advantage of DB-API's implicit transactions, it
allows multiple queries to be issued in sequence on the same connection.