[archive] Handle insert and update more compactly
The previous approach to adding a bookmark involved figuring out the
existing state of a given within the database, and then figuring out
an appropriate course of action. The new approach pushes all of that
to the database, leveraging sqlite's changes() function to insert
when an update has not occurred.