[later] Rework tests to account for markup plugin
Tests that were previously able to make unmocked requests and inspect
the response body can no longer do so, because the markup plugin is
not subscribed during the test. Switch these tests to mocks.

Rework the markup plugin so that the plain text parser is more similar
to the other parser classes.