Reorient provisioning process around rsync
Discontinue previous approach of doing a git pull from Github, which
was annoying, in favor of rsync against the current state of the
checkout. Consolidate ansible roles to decrease sprawl. Add make
targets for invoking vagrant more conveniently. Reorient example
config around a production setup, since this is the file that is
populated when provisioning.