Switch back to simple audio for wav file playback
With the introduction of the scheduler plugin, playing audio through
mpd is no longer compelling. The mpd setup also tended to not work
when playing a newly-generated file for the first time, probably
because of a timing issue between receiving the wav file and adding it
to the mpd database.

The file caching setup remains the same, only the playback mechanism
has changed.