Add cache-bust jinja filter
Cache-busting appends an identifier to URLs so that things stay
up-to-date when caching headers are used.

The jinja filter uses the contextfilter decorator so that it always
gets called. Without it, bytecode caching would cause the identifiers
to be calculated once and never recalculated until the next template
change.

This is mainly a benefit during development. To avoid unnecessary
identifier recalculation in production, use a server configuration
value to determine whether identifiers should be cached in memory via
the memorize plugin.