[redirect] Add redirect app
This is a simplistic app for referer privacy. Given a URL via
querstring, it will return an HTML page that performs a client-side
redirect as well as specify a same-origin referrer policy. By doing
so, it's less likely that the remote site will see a referer on the
incoming request.

This replaces an earlier, standalone project in PHP called ubounce.