[redirect] Standardize on u querystring parameter
Rather than allow URLs to use provided as the entirety of the
querystring, require them to be provided under the key "u" for less
ambiguity. Although unlikely to be needed, this would allow other
values to be specified in the querystring whereas the other approach
does not.

Also run the JS redirect as soon as possible to reduce lag.