Allow visitors queries to be cached at database
Add additional logic to the logindex plugin and sqlite mixins so that
stored visitors queries can be pre-cached. The existing cache plugin
is not used for this because sqlite3.Row objects resisted
serialization. It would have also been a hassle to incorporate them
into the controller. The alternative is to copy rows from the main
logs table into dedicated cache tables, and then divert queries to
them accordingly.