[logindex] Only parse domain from user agent
Remove the ua-parser library and change the logindex schema to agent
and agent_domain fields. There isn't enough value to be had from
parsing the user agent--the results aren't always reliable, the agent
strings aren't always trustworthy, and the value of seeing browser and
platform version isn't that substantial for my purposes.

The agent field returns to the schema so that it is in easy reach for
future parsing or analysis, but is otherwise not displayed in the
visitors app.