Drop monit role from ansible config
Monit was originally used to auto restart the server, but was never
upgraded when the startup script was migrated to systemd. Since
systemd can otherwise take care of this, and since the other monit
task related to watching /var/log/messages never proved to be useful,
drop the whole thing. Instead, upgrade the systemd service with
restart logic.