Bring back local gzipping
Gzip was formerly disabled so that a reverse proxy like Nginx would
take care of it. But this complicates the use of Etags, because
compression invalidates the Etag--the resource has changed, so the tag
is presumed invalid. Using a weak etag could work, but involves extra
server configuration. Handling gzip locally lets the etag through and
seems easier all around.