[jenkins] Reference branch name in notification
Rather than relying on the branch to be part of the build name, as
was case for customized payloads triggered from pipeline builds, add
it to the notification title explicitly.

This makes it possible for stage and production notifications to use
the same local ID, and for one to auto-clear the other. Previously, a
stage build followed by a production build would result in two
notifications instead of the production notification automatically
clearing whatever had come before.

Also, the distinction between pipeline-generated payloads and plugin
payloads is removed because their structure is no longer different.