[logindex] Drop expression-based index
Add the datestamp field to accomplish what the previous
expression-based index was otherwise doing automatically. As before
with the discrete year, month, and day fields, calculate and populate
the datestamp during parsing.

The motivation for this denormalization is sqlite versioning. The
current version of sqlite on Debian Stretch is 3.16.2, which treats
date/time functions like the one used here as
non-deterministic. Non-deterministic expressions can't be used on
expression-based indexes prior to sqlite 3.20.0.