[logindex] Drop query precaching
It was meant to improve the responsiveness of the visitors app by
duplicating the main table into subsets that could be queried more
quickly. Recent improvements to table indexes have made this not as
much of a problem as it was, and the extra work from having the caches
replenished after every indexing run is otherwise wasting resources.