[startpage] Be more strict about canonical URLs
In preparation for implementing service workers, be more strict and
less wiki-esque about URLs for the default page are handled. Redirect
to the canonical URL at all times.

Also add special handling for the service worker script to ensure its
scope is application-wide and not limited to the static directory.