[visitors] Map cookie values to IP for more reliable display
If the newest logline for an IP was for a URL that didn't include the
cookie field, the cookie field of any subsequent loglines would not be
shown. Prevent this by creating a mapping between IP addresses and
cookie values, and using that as the basis for displaying the "Search
for visits with this cookie" icon rather the logline itself.