[registry] Replace copy-key toolbar button with new-item button
Copying the key of an existing record is usually done in the context
of creating another record with the same key. The new-item button
redirects to the add view and pre-populates the key field to
accomplish this in one click, instead of copying the key then
switching to the add view then pasting.