[speak] Allow pronunciation customizations
Store search-and-replace pairs as part of the application config, and
apply them to the statement to be spoken prior to assembling the SSML
document.

The SSML document is now handled as a string template rather than
being built with minidom to make it easier to include tags in the
replacement values.