Detect file extension mime override more explicitly
Just matching on . in the path involved an implicit assumption that
clashed with the use of IP addresses in urls for the whois app. This
can be avoided by restricting the match to the start of the
string. It's viable here because there's no concept of a default
document, so something like "/index.txt" doesnn't have to be
considered.