Shift default configuration to production-friendly values
The out-of-the-box configuration for the server is now geared toward
production rather than development so that there is less need for
supplemental configuration. This makes the deployment process slightly
easier.

Part of this change involves removing file-based application and
access logs, which were redundant with the logging that is already
happening through systemd and nginx. Additional configuration logic
was added to get CherryPy not to write access logs to stdout.