[bookmarks] Make default view more performant
Limit the list of recent bookmarks to the past 6 months so that sqlite
can avoid a full table scan. The new added_date field is redundant
with the existing added field, but is being introduced anyway as a
result of the production version of sqlite not being able to create an
expression-based index using the date function. In the future when a
newer version of sqlite is available it should be possible to drop
this redundant field.