Use cache.add() rather than cache.put()
The two methods are similar, but add() handles the fetch as well as
the storage. Also perform a delete prior to the add so that existing
entries can be cleaned up. Unlike delete() and match(), add() doesn't
take an options parameter so querystring becomes part of the cache key
which is not otherwise helpful here.