Revert to individual imports of cherrypy plugins and tools
Automatic importing via importlib from __init__.py wasn't satisfying
because it made things less obvious. To understand how a plugin was
imported, you had to know to look in the init file and then understand
what it was doing and why the special handling it was doing for
certain files was necessary. Although individual imports are more
verbose, they are also more straightforward.