Add mypy and start adding type hints
As a first step, go back to definining exposed and user_facing in all
app controllers rather than using importlib in the server to set
defaults. This is more repetitive, but sidesteps some typing
complications. Although importlib was valued at the time it was
implemented for cutting down on repetition, I'm more interested in
having things be more obvious and less magical.