Queue applog messages instead of writing to database directly
The previous approach of using a persistent database connection shared
among threads was still susceptible to occasional
sqlite3.OperationalError exceptions ("cannot start a transaction
within a transaction"). I also didn't like having the -shm and -wal
files always hanging around. Instead do as is done in logindex and put
messages in a deque which is then scheduled for processing.