[bookmarklets] Open bounce bookmarklet in current window
The browser might leave the bookmarklet JS behind in the calling
page's address bar if a new window is opened. Updating the location of
the current page avoids this. The new window behavior is also not
otherwise obvious at first, and you might not always want to stay on
the current page. But if you do, the browser probably provides a tab
cloning feature.