[reminder] Handle revived reminders without crashing the scheduler
Use of a named tuple caused problems if the server restarted before
the event fired because the tuple datastructure was lost during
serialization. Serialize a dict instead and handle this specially
during revive.

Also change the deletion notification to a retraction so that the UI
can be consistent. Previously it acted as if the delete button had
been clicked when a timer ran out, resulting in an innacurate
notification. The cancellation message wasn't that useful anyway.