Disable CherryPy encode tool and normalize controller signatures
Since calling encode() is necessary for JSON responses, relying on the
encode tool for HTML responses isn't as obvious as handling it
explicitly. The impact to controllers is minimal since they are
already returning the result of jinja:render. But this allows for a
single return type rather than a union of string or bytes.