[parse] Define grammars as needed rather than at startup.
The pyparsing grammars were previously generated at plugin startup,
which is wasteful given how infrequently they are used. Skip that work
until the grammar is needed. For logindex queries this will still be
wasteful in that the grammar is recreated, but it's better than having
startup time and memory usage be impacted unnecessarily.