Don't re-run schema creation queries at startup.
This was originally meant as a way to pick up schema changes
automatically, but in practice there is rarely a need for it. When
schemas do change, it's usually a larger change that wouldn't work
with this setup anyway. In the typical case the server is doing
unnecessary work at every startup, so just reduce that to a
file-exists check instead.