[assets] Add plugin for serving static files from sqlite.
This is a reaction to bad IO performance when the application runs as
a zipapp and tries to serve static assets. Too much IO was needed to
both hash assets for cache busting and to serve requests. Offloading
this work to sqlite allows for easier access and room to precalculate
file hashes.