[scheduler] Delay revive until cache is ready.
When the cache database is being created, the scheduler may try to
query it too early since both it and the cache plugin are listening on
the server:ready channel. When this happens, there is an error about
the unexpired view not existing. Having the scheduler instead listen
for a new new cache:ready before attempting to revive avoids this.