Give more time to startup in systemd service.
The default value seems to be 1min 30sec, which may not be enough if
the host has a lot of things going on at system boot. Actual startup
time of the application isn't amazing, but also isn't terrible. This
is more of a quick-fix in lieu of something more definitive from the
systemd side.