Use scripts for HTTP testing instead of orgmode code blocks.
Having everything in urls.org ended up being a hassle. Although well
organized, there was too much syntax and noise from all the setup
needed for each request. The alternative is to have a collection of
Python scripts that make HTTP requests via httpx and can be invoked
from Makefile and can be made as simple or complicated as needed.